Use searchable for tag list and force refresh tags using refreshable
Closes #23
